added PC8801 driver
A driver for the PC8801 was added, however is experimental and has not been tested. The layer logic was completely reprogrammed since the previous logic was only able to toggle between layers. The driver and keyboard selection is now file based, reading the KEYBOARD variable from the Makefile to get the information.